Add padded individual links to prevent spoilers
The displayed individual links are padded to be the same length in order
to prevent the gift exchange organizer from being spoiled of the
pairings. These links are advanced individual links of type 0:
everything after the name is just random bytes.

The copy and share buttons use the basic individual links.
The copy event in the input link elements is handled so that the basic
individual links are copied instead.